I managed to get my handleSubmit and handleChange working
I struggled for a really long time because it said I was passing an object as a React child and I couldn't for the life of me find where I did that.
Took me literally hours to realize it was because I had my myFriend object from state down in the render because I forgot it would be captured when the mapping happened.
Also took me entirely too long to realize that all of the different parts of state were updating the same way at every change, which was messing up my keys and messing up my renders.