Add ability to update a friend
I went through and completed the code needed to allow the user to update the information about a specific friend. There's some optimization that needs to be finished, but the base functionality is there. All I need to do is complete the ability to delete a specific friend.  That should be pretty simple.